Flesh out TZif file parsing
Clean up the code that parses a TZif file and extend struct time_zone
to represent all fields.  Note that we've made a conscious decision to
ignore the v1 data; version 2 was introduced in 2005 and is unlikely
to be used anywhere that might benefit from this implementation.

Also, adjust tzdump to print this "cooked" information by
default (with a -r flag to the data directly from the file).