Have tzalloc search more places for a time zone.
If its parameter is NULL, do localtime; if it's the empty string, do
UTC, if it begins with a colon then it's an absolute path (if it
begins with a slash) or a relative path (otherwise).  If it doesn't
begin with a colon then try treating it as a path or, failing that, as
a TZ string.  We don't yet support this last option.