Do a better job of handling times in gaps
Users should be able to take the current time, convert it to a struct
tm with localtime_rz, add or subtract some number of hours, days, etc,
and convert that back to a timestamp with mktime_z to get a canonical
struct tm.  Fix the handling of that when the manipulated time falls
into the gap created by a spring forward.