Avoid some unnecessary re-canonicalizations
The canonicalized tm will be correct in many common use cases, so
let's not waste cylces recalculating it unless (1) the time after
canonicalization is is not a real time (i.e. it falls into a gap
between standard time and daylight savings) or (2) the time zone
supports leap seconds and the supplied number of seconds is not
between 0 and 59.