Have a stab at using the TZ string
The time zone files and with a TZ string that we can use to calculate
when DST starts and ends after the last explicit transition.  Use that
to calculate the transition times a block of 400 years and just index
into that to find the offset to use.  Only the forward translation is
supported thus far.