Further decrease per-timezone memory usage
There are only 14 different types of years -- a set of 7 for the first
day of the year for a normal year, and another set of 7 for leap
years.  Each year of a given type will have its transitions the same
number of seconds since the beginning of the year, so we only need 14
start offsets and 14 end offsets for any given TZ string.