Use division rather than bisecting.
When looking up the start of the year, we can divide by the average
number of seconds per year to get a lower bound on the index, and just
step forward to find the match.  We should never have to check more
than 3 timestamps to find the one we're looking for; this is
noticeably faster than bisecting.