Improve test-mktime error reporting
When test-mktime's canonicalization fails, it would be handy if it
reported the expected and actual struct tm information, like
test-endpoints does.  Let's shift assert_tm_eq into a common location,
and let's split it into a comparison part (which should be fast for
test-endpoints' exhaustive mode) and a reporting part (which need not
be optimized), and extend it to report all fields.