Fix the calculation of the adjusted timestamp
When looking up the time zone offset to use as specified by a POSIX TZ
string, we use the modulus opeartor to fold the timestamp into a
400-year range.  If the timestamp is less than the reference time,
then the result of that modulus will be negative, and we need to add
secs_per_400_years to get into the desired range.  Let's shift that
logic into a separate function and use it where it's needed.