Teach tz64_tm_to_ts to cheat
In cases where the DST indicator is ambiguous we can try to get some
clarity from gmtoff.  That can bias the results for, say, London,
where an zero is both a valid gmtoff and also a common arbitrary value
tha could be assigned to the field, but on the whole it seems like a
good idea: it makes the function a reliable inverse of tz64_ts_to_tm.