Try to handle leap seconds with unusual UTC offsets
There's nothing to prevent a government from choosing to adopt an
offset from UTC that's not a whole number of minutes.  If that were to
happen, then any subsequent leap seconds wouldn't happen at the end of
the minute.  Presumably we'd want to still count from 0 through 60 of
the minute in which the leap second occurred, so let's try to do that.