Add a work-around for redundant final timestamps
While glibc uses the POSIX TZ string for timestamps after the last
explicit entry in the TZif file, performance tests show that it is
substantially slower to convert time in this range.  Perhaps in
response to that, glibc's zic implementation appends a redundant
timestamp for INT32_MAX.  Our conversion functions are slightly faster
after the last transition for time zones that don't observe daylight
savings time, so let's detect and remove the redundant entries.