Use Tony Finch's math to canonicalize year, month and day
The approach of finding a 400 year block relative to 2001 and
repeatedly subtracting 1, 4, 20 or 100 years at a time from it
does work, it can be quite slow in the worst case.  Thankfully,
Tony Finch has worked out a set of simple operations to convert
year, month, day to a day count and back.  Use those instead
so that the best and worst case times are more closely aligned.