Add some test cases for before year 1
Even though the Gregorian calendar wasn't in use then, we can
still calculate the what date a given time stamp corresponds to
if we extend the calendar back in time.  Do that, both to be
consistent with other time conversion implementations and a because
it's more useful than returning garbage.